new_backend/followup_scheduler/requirements.txt
# This file specifies the dependencies for the Azure Function.
# It should be deployed alongside the function code.

azure-functions
sqlalchemy
psycopg2-binary
sendgrid 